Move aws helpers to separate package
