+++ common/aws/package.json
@@ -1,0 +1,16 @@
+{
+    "name": "@casimir/aws",
+    "private": true,
+    "main": "src/index.ts",
+    "scripts": {
+        "build": "echo '@casimir/aws build not specified. Disregard this warning and any listed errors above if @casimir/aws is not needed for the current project build.' && exit 0",
+        "test": "echo \"Error: no test specified\" && exit 1"
+    },
+    "dependencies": {
+        "@aws-sdk/client-secrets-manager": "^3.204.0",
+        "@aws-sdk/credential-providers": "^3.204.0"
+    },
+    "devDependencies": {
+        "@types/node": "^17.0.38"
+    }
+}
