+++ common/helpers/package.json
@@ -7,8 +7,6 @@
         "test": "echo \"Error: no test specified\" && exit 1"
     },
     "dependencies": {
-        "@aws-sdk/client-secrets-manager": "^3.204.0",
-        "@aws-sdk/credential-providers": "^3.204.0",
         "ethers": "^5.7.2"
     },
     "devDependencies": {
